renames PQC class
